Gestione errori nelle API

+++ app/orders.js
@@ -18,27 +18,48 @@ router.post("", tokenChecker, async (req, res) => {
     const { customerId, status } = req.body;
     let data = null;
     let cart = await Cart.findOne({ userId: customerId });
-    let customer = await Cliente.findById(customerId).populate(
-      "shipping_address"
-    );
-    const product = await Product.findById(cart.items[0].productId);
-    const orderData = {
-      customerId: customerId,
-      items: cart.items,
-      subTotal: cart.subTotal,
-      sellerId: product.userId,
-      status: status,
-      payment: customer.payment,
-      shipping_address: customer.shipping_address,
-    };
-    let order = new Order(orderData);
-    data = await order.save();
-    console.log(order);
-    return res.status(200).send({
-      code: 200,
-      message: "Order created successfully!",
-      data: data,
-    });
+    if (cart) {
+      let customer = await Cliente.findById(customerId).populate(
+        "shipping_address"
+      );
+      if (customer) {
+        const product = await Product.findById(cart.items[0].productId);
+        if (product) {
+          const orderData = {
+            customerId: customerId,
+            items: cart.items,
+            subTotal: cart.subTotal,
+            sellerId: product.userId,
+            status: status,
+            payment: customer.payment,
+            shipping_address: customer.shipping_address,
+          };
+          let order = new Order(orderData);
+          data = await order.save();
+          console.log(order);
+          return res.status(200).send({
+            code: 200,
+            message: "Order created successfully!",
+            data: data,
+          });
+        } else {
+          res.status(401).json({
+            success: false,
+            message: "Errore nel GET del prodotto",
+          });
+        }
+      } else {
+        res.status(402).json({
+          success: false,
+          message: "Errore nel GET del cliente",
+        });
+      }
+    } else {
+      res.status(403).json({
+        success: false,
+        message: "Errore nel GET del carrello dell'utente",
+      });
+    }
   }
 });
 
@@ -53,17 +74,24 @@ router.get("/:id", tokenChecker, async (req, res) => {
     let order = await Order.findById(req.params.id).populate(
       "shipping_address"
     );
-    res.status(200).json({
-      self: "/api/v2/orders/" + order.id,
-      customerId: order.customerId,
-      items: order.items,
-      subTotal: order.subTotal,
-      status: order.status,
-      id: order.id,
-      payment: order.payment,
-      sellerId: order.sellerId,
-      shipping_address: order.shipping_address,
-    });
+    if (order) {
+      res.status(200).json({
+        self: "/api/v2/orders/" + order.id,
+        customerId: order.customerId,
+        items: order.items,
+        subTotal: order.subTotal,
+        status: order.status,
+        id: order.id,
+        payment: order.payment,
+        sellerId: order.sellerId,
+        shipping_address: order.shipping_address,
+      });
+    } else {
+      res.status(401).json({
+        success: false,
+        message: "Errore nel GET del singolo ordine",
+      });
+    }
   }
 });
 
@@ -72,20 +100,27 @@ router.get("/catalog/:id", tokenChecker, async (req, res) => {
 
   if (user_type === "Cliente" || user_type === "Admin") {
     let orders = await Order.find({ customerId: req.params.id });
-    orders = orders.map((order) => {
-      return {
-        self: "/api/v2/orders/" + order.id,
-        customerId: order.customerId,
-        items: order.items,
-        subTotal: order.subTotal,
-        status: order.status,
-        id: order.id,
-        payment: order.payment,
-        sellerId: order.sellerId,
-        shipping_address: order.shipping_address,
-      };
-    });
-    res.status(200).json(orders);
+    if (orders) {
+      orders = orders.map((order) => {
+        return {
+          self: "/api/v2/orders/" + order.id,
+          customerId: order.customerId,
+          items: order.items,
+          subTotal: order.subTotal,
+          status: order.status,
+          id: order.id,
+          payment: order.payment,
+          sellerId: order.sellerId,
+          shipping_address: order.shipping_address,
+        };
+      });
+      res.status(200).json(orders);
+    } else {
+      res.status(401).json({
+        success: false,
+        message: "Errore nel GET degli ordini del cliente",
+      });
+    }
   }
 });
 
@@ -94,20 +129,27 @@ router.get("/catalogv/:id", tokenChecker, async (req, res) => {
 
   if (user_type === "Venditore" || user_type === "Admin") {
     let orders = await Order.find({ sellerId: req.params.id });
-    orders = orders.map((order) => {
-      return {
-        self: "/api/v2/orders/" + order.id,
-        customerId: order.customerId,
-        items: order.items,
-        subTotal: order.subTotal,
-        status: order.status,
-        id: order.id,
-        payment: order.payment,
-        sellerId: order.sellerId,
-        shipping_address: order.shipping_address,
-      };
-    });
-    res.status(200).json(orders);
+    if (orders) {
+      orders = orders.map((order) => {
+        return {
+          self: "/api/v2/orders/" + order.id,
+          customerId: order.customerId,
+          items: order.items,
+          subTotal: order.subTotal,
+          status: order.status,
+          id: order.id,
+          payment: order.payment,
+          sellerId: order.sellerId,
+          shipping_address: order.shipping_address,
+        };
+      });
+      res.status(200).json(orders);
+    } else {
+      res.status(401).json({
+        success: false,
+        message: "Errore nel GET degli ordini del venditore",
+      });
+    }
   }
 });
 
@@ -120,6 +162,12 @@ router.delete("/:id", tokenChecker, async (req, res) => {
     user_type === "Venditore"
   ) {
     let order = await Order.findById(req.params.id).exec();
+
+    if (!order) {
+      res.status(401).send();
+      console.log("Order not found");
+      return;
+    }
 
     await order.deleteOne();
 
@@ -135,12 +183,19 @@ router.put("/:id", tokenChecker, async (req, res) => {
     let order = await Order.findByIdAndUpdate(req.params.id, {
       status: req.body.status,
     });
-    let orderId = order.id;
-    console.log("Stato dell'ordine modificato correttamente");
-    res
-      .location("/api/v2/orders/" + orderId)
-      .status(200)
-      .send();
+    if (order) {
+      let orderId = order.id;
+      console.log("Stato dell'ordine modificato correttamente");
+      res
+        .location("/api/v2/orders/" + orderId)
+        .status(200)
+        .send();
+    } else {
+      res.status(401).json({
+        success: false,
+        message: "Errore nel PUT dell'ordine",
+      });
+    }
   }
 });
